Added error handling

+++ controller/productController.js
@@ -1,41 +1,47 @@
 const Product = require("../models/product");
+const statusUtils = require("../utilities/statusUtils");
+const mongoose = require("mongoose");
 
-const PRODUCT_NOT_FOUND = {
-  message: "No Product Found",
-  status: "NO_PRODUCT_FOUND",
-};
-const PRODUCT_DELETED = {
-  message: "Product Successfully Deleted",
-  status: "PRODUCT_DELETED",
-};
-const SERVER_ERROR = {
-  message: "Something went wrong",
-  status: "SERVER_ERROR",
-};
-
-const getAllProducts = async (_, res) => {
+const { getProductNotFoundError, getServerError, PRODUCT_DELETED_MSG } =
+  statusUtils;
+const getAllProducts = async (_, res, next) => {
   try {
     const products = await Product.find();
     return res.json(products);
   } catch (err) {
     console.error(err);
-    return res.status(500).json(SERVER_ERROR);
+    err = getServerError();
+    /**here since we are not inside any promise code, throw will also work intead of next(error)
+     *throw err
+     if we are throwing an error inside promise then catch,there it will not work as it is running asynchronously. There you have to use next(error)
+     */
+    next(err);
   }
 };
 
-const getProductById = async (req, res) => {
+const getProductById = async (req, res, next) => {
   try {
     const prdId = req.params.id;
+    if (!mongoose.Types.ObjectId.isValid(prdId)) {
+      const err = getProductNotFoundError();
+      throw err;
+    }
     const product = await Product.findById(prdId);
-    if (!product) return res.status(404).json(PRODUCT_NOT_FOUND);
+    if (!product) {
+      const err = getProductNotFoundError();
+      throw err;
+    }
     return res.json(product);
   } catch (err) {
     console.error(err);
-    return res.status(500).json(SERVER_ERROR);
+    if (!err.statusCode) {
+      err = getServerError();
+    }
+    next(err);
   }
 };
 
-const addProduct = async (req, res) => {
+const addProduct = async (req, res, next) => {
   try {
     const product = await Product.create({
       title: req.body.title,
@@ -44,19 +50,24 @@ const addProduct = async (req, res) => {
       imgUrl: req.body.imgUrl,
       description: req.body.description,
     });
-
     res.status(201).json(product);
   } catch (err) {
     console.error(err);
-    res.status(500).json(SERVER_ERROR);
+    if (!err.statusCode) {
+      err = getServerError();
+    }
+    next(err);
   }
 };
 
-const updateProduct = async (req, res) => {
+const updateProduct = async (req, res, next) => {
   try {
     const prdId = req.params.id;
     let product = await Product.findById(prdId);
-    if (!product) return res.status(404).json(PRODUCT_NOT_FOUND);
+    if (!product) {
+      const err = getProductNotFoundError();
+      throw err;
+    }
     product.title = req.body.title;
     product.price = req.body.price;
     product.qty = req.body.qty;
@@ -66,19 +77,32 @@ const updateProduct = async (req, res) => {
     res.json(product);
   } catch (err) {
     console.error(err);
-    res.status(500).json(SERVER_ERROR);
+    if (!err.statusCode) {
+      err = getServerError();
+    }
+    throw err;
   }
 };
 
-const deleteProduct = async (req, res) => {
+const deleteProduct = async (req, res, next) => {
   try {
     const prdId = req.params.id;
+    if (!mongoose.Types.ObjectId.isValid(prdId)) {
+      const err = getProductNotFoundError();
+      throw err;
+    }
     const product = await Product.findByIdAndDelete(prdId);
-    if (!product) res.status(404).json(PRODUCT_NOT_FOUND);
-    return res.json(PRODUCT_DELETED);
+    if (!product) {
+      const err = getProductNotFoundError();
+      throw err;
+    }
+    return res.json(PRODUCT_DELETED_MSG);
   } catch (error) {
     console.error(error);
-    res.status(500).json(SERVER_ERROR);
+    if (!error.statusCode) {
+      error = getServerError();
+    }
+    next(error);
   }
 };
